refactor: split extension utils/types.ts into three owner modules
The Settings interface joins settings/storage.ts, the popup-to-content-script
protocol (PING, MANUAL_SPACING, ContentScriptResponse, CONTENT_SCRIPT_LOADED)
becomes src/messages.ts beside the entry files it joins, and the classify
protocol (Candidate, ClassifyCandidates*, ClassifiedCandidate, CandidateLabel)
becomes ai-spacing/messages.ts. utils/ is now gone. No behavior change.
