+++ browser-extensions/chrome/src/ai-spacing/messages.ts
@@ -1,39 +1,4 @@
-import type { HyphenLabel } from '../ai-spacing/hyphen-prompt';
-
-export interface Settings {
-  spacing_mode: 'spacing_when_load' | 'spacing_when_click';
-  filter_mode: 'blacklist' | 'whitelist';
-  blacklist: string[];
-  whitelist: string[];
-  is_mute_sound_effects: boolean;
-  is_enable_text_autospace: boolean;
-  is_enable_ai_spacing: boolean;
-}
-
-export interface PingMessage {
-  action: 'PING';
-}
-
-export interface ManualSpacingMessage {
-  action: 'MANUAL_SPACING';
-}
-
-// Messages sent TO content script (via chrome.tabs.sendMessage)
-export type MessageToContentScript = PingMessage | ManualSpacingMessage;
-
-// We only need a response when sender actually needs it, e.g.,
-// popup needs to know if content script is loaded or not,
-// or if manual spacing is successful or not
-export interface ContentScriptResponse {
-  success: boolean;
-}
-
-export interface ContentScriptLoadedMessage {
-  type: 'CONTENT_SCRIPT_LOADED';
-}
-
-// Messages sent FROM content script to extension (via chrome.runtime.sendMessage)
-export type MessageFromContentScript = ContentScriptLoadedMessage;
+import type { HyphenLabel } from './hyphen-prompt';
 
 // One flagged symbol. `at` is its index inside `sentence` rather than a slice of its own, because the symbol's own characters do not identify which one is meant when a sentence carries a second one.
 export interface Candidate {
